+++ src/main/java/controller/usuarios/BorrarUsuarioServlet.java
@@ -36,7 +36,7 @@ public class BorrarUsuarioServlet extends HttpServlet {
 
 	@Override
 	protected void doPost(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-		Long id = Long.parseLong(req.getParameter("id"));
+		Integer id = Integer.parseInt(req.getParameter("id"));
 		
 		try {
 			usuarioService.delete(id);
